+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 
 ## Unreleased
+
+
+## 1.0.0
 ### Added
 - tests for python 3.7 and 3.8
 ### Changed